refactor(schemas): rename general-settings/ → site-settings/ to match table name
The directory holds the `site_settings` table. Renaming for consistency
between table name and directory name. JSDoc `@module` paths and consumer
imports updated to match. Non-schema files (api/general-settings.ts,
services/general-settings/) retain their names — out of scope here.
