+++ packages/nextly/src/schemas/site-settings/postgres.ts
@@ -1,18 +1,18 @@
 /**
- * MySQL Schema for General Settings
+ * PostgreSQL Schema for General Settings
  *
- * Defines the `site_settings` singleton table for MySQL.
+ * Defines the `site_settings` singleton table for PostgreSQL.
  * Only one row ever exists (id = 'default').
  *
- * @module schemas/general-settings/mysql
+ * @module schemas/site-settings/postgres
  * @since 1.0.0
  */
 
-import { mysqlTable, varchar, datetime } from "drizzle-orm/mysql-core";
+import { pgTable, text, varchar, timestamp } from "drizzle-orm/pg-core";
 
-export const siteSettingsMysql = mysqlTable("site_settings", {
+export const siteSettingsPg = pgTable("site_settings", {
   /** Always 'default' — enforces singleton pattern. */
-  id: varchar("id", { length: 50 }).primaryKey(),
+  id: text("id").primaryKey(),
 
   /** Display name for the application. */
   applicationName: varchar("application_name", { length: 255 }),
@@ -39,14 +39,16 @@ export const siteSettingsMysql = mysqlTable("site_settings", {
   logoUrlDark: varchar("logo_url_dark", { length: 2048 }),
 
   /** JSON array of custom sidebar groups, e.g. [{"slug":"analytics","name":"Analytics","icon":"BarChart"}] */
-  customSidebarGroups: varchar("custom_sidebar_groups", { length: 4096 }),
+  customSidebarGroups: text("custom_sidebar_groups"),
 
   /** JSON object mapping plugin slugs to sidebar group overrides, e.g. {"form-builder":"collections"} */
-  pluginPlacements: varchar("plugin_placements", { length: 4096 }),
+  pluginPlacements: text("plugin_placements"),
 
   /** When the settings were last updated. */
-  updatedAt: datetime("updated_at").notNull(),
+  updatedAt: timestamp("updated_at", { withTimezone: false })
+    .defaultNow()
+    .notNull(),
 });
 
-export type SiteSettingsMysql = typeof siteSettingsMysql.$inferSelect;
-export type SiteSettingsInsertMysql = typeof siteSettingsMysql.$inferInsert;
+export type SiteSettingsPg = typeof siteSettingsPg.$inferSelect;
+export type SiteSettingsInsertPg = typeof siteSettingsPg.$inferInsert;
